Move kafka code to separate file
